requirements.txt
transmissionrpc==0.11
pytest==3.6.4
grpcio-tools==1.26.0